Refactor pollers into separate modules
